# ironflow-sdk/openapi.json
{
  "components": {
    "schemas": {
...
      "AccountKindResponse": {
        "description": "A supported account kind.",
        "properties": {
          "display_name": {"description": "Human-readable name.", "type": "string"},
          "fields": {"description": "Fields of the add form.", "items": {"$ref": "#/components/schemas/AccountFormFieldResponse"}, "type": "array"},
          "id": {"description": "Kind identifier.", "type": "string"}
        },
        "required": ["id", "display_name", "fields"],
        "type": "object"
      },
...
      "AccountTestResult": {
        "description": "Outcome of a live credential test.",
        "enum": ["valid", "limited", "unauthorized"],
        "type": "string"
      },
...
      "ProviderAccountTestResponse": {
        "description": "Response of `POST /provider-accounts/{id}/test`.",
        "properties": {
          "result": {"$ref": "#/components/schemas/AccountTestResult"},
          "windows": {"description": "Windows reported by the provider.", "items": {"$ref": "#/components/schemas/AccountWindowResponse"}, "type": "array"}
        },
        "required": ["result", "windows"],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "name": "Thomas Tartrau"
    },
    "description": "REST API for the ironflow workflow engine",
    "license": {
      "identifier": "MIT",
      "name": "MIT"
    },
    "title": "Ironflow REST API",
    "version": "1.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
    "/api/v1/provider-accounts": {
      "get": {
        "description": "The response never includes the credential.",
        "operationId": "list_provider_accounts",
        "parameters": [
          {"description": "Only accounts of this kind.", "in": "query", "name": "kind", "required": false, "schema": {"type": "string"}},
          {"description": "Page number (1-based).", "in": "query", "name": "page", "required": false, "schema": {"format": "int32", "minimum": 0, "type": "integer"}},
          {"description": "Items per page (max 100).", "in": "query", "name": "per_page", "required": false, "schema": {"format": "int32", "minimum": 0, "type": "integer"}}
        ],
        "responses": {
          "200": {"content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/ProviderAccountResponse"}, "type": "array"}}}, "description": "Provider accounts listed"},
          "401": {"description": "Unauthorized"},
          "403": {"description": "Forbidden"}
        },
        "security": [{"Bearer": []}],
        "summary": "List Provider Accounts with their latest windows. Admin only.",
        "tags": ["provider-accounts"]
      },
      "post": {
        "description": "The token is checked against the provider before anything is stored.",
        "operationId": "create_provider_account",
        "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/CreateProviderAccountRequest"}}}, "description": "Account and its credential", "required": true},
        "responses": {
          "201": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProviderAccountResponse"}}}, "description": "Provider account created"},
          "400": {"description": "Invalid input or unknown kind"},
          "401": {"description": "Unauthorized"},
          "403": {"description": "Forbidden"},
          "409": {"description": "Name already taken"},
          "422": {"description": "Credential rejected"},
          "502": {"description": "Provider unreachable"}
        },
        "security": [{"Bearer": []}],
        "summary": "Add a Provider Account. Admin only.",
        "tags": ["provider-accounts"]
      }
    },
    "/api/v1/provider-accounts/kinds": {
      "get": {
        "operationId": "list_account_kinds",
        "responses": {
          "200": {"content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/AccountKindResponse"}, "type": "array"}}}, "description": "Supported kinds"},
          "401": {"description": "Unauthorized"},
          "403": {"description": "Forbidden"}
        },
        "security": [{"Bearer": []}],
        "summary": "List the account kinds this server supports, with their form fields. Admin only.",
        "tags": ["provider-accounts"]
      }
    },
    "/api/v1/provider-accounts/{id}": {
      "delete": {
        "operationId": "delete_provider_account",
        "parameters": [{"description": "Account UUID or name", "in": "path", "name": "id", "required": true, "schema": {"type": "string"}}],
        "responses": {
          "204": {"description": "Provider account deleted"},
          "401": {"description": "Unauthorized"},
          "403": {"description": "Forbidden"},
          "404": {"description": "Provider account not found"}
        },
        "security": [{"Bearer": []}],
        "summary": "Delete a Provider Account and its credential. Admin only.",
        "tags": ["provider-accounts"]
      },
      "get": {
        "operationId": "get_provider_account",
        "parameters": [{"description": "Account UUID or name", "in": "path", "name": "id", "required": true, "schema": {"type": "string"}}],
        "responses": {
          "200": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProviderAccountResponse"}}}, "description": "Provider account"},
          "401": {"description": "Unauthorized"},
          "403": {"description": "Forbidden"},
          "404": {"description": "Provider account not found"}
        },
        "security": [{"Bearer": []}],
        "summary": "Get a Provider Account by UUID or name. Admin only.",
        "tags": ["provider-accounts"]
      },
      "patch": {
        "operationId": "update_provider_account",
        "parameters": [{"description": "Account UUID or name", "in": "path", "name": "id", "required": true, "schema": {"type": "string"}}],
        "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/UpdateProviderAccountRequest"}}}, "description": "Fields to change", "required": true},
        "responses": {
          "200": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProviderAccountResponse"}}}, "description": "Provider account updated"},
          "400": {"description": "Invalid input"},
          "401": {"description": "Unauthorized"},
          "403": {"description": "Forbidden"},
          "404": {"description": "Provider account not found"},
          "422": {"description": "Credential rejected"},
          "502": {"description": "Provider unreachable"}
        },
        "security": [{"Bearer": []}],
        "summary": "Update a Provider Account. Admin only.",
        "tags": ["provider-accounts"]
      }
    },
    "/api/v1/provider-accounts/{id}/test": {
      "post": {
        "operationId": "test_provider_account",
        "parameters": [{"description": "Account UUID or name", "in": "path", "name": "id", "required": true, "schema": {"type": "string"}}],
        "responses": {
          "200": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProviderAccountTestResponse"}}}, "description": "Credential checked"},
          "401": {"description": "Unauthorized"},
          "403": {"description": "Forbidden"},
          "404": {"description": "Provider account not found"},
          "422": {"description": "No credential stored"},
          "502": {"description": "Provider unreachable"}
        },
        "security": [{"Bearer": []}],
        "summary": "Check the stored credential against the provider. Admin only.",
        "tags": ["provider-accounts"]
      }
    },
    "/api/v1/provider-accounts/{id}/usage": {
      "get": {
        "operationId": "provider_account_usage",
        "parameters": [
          {"description": "Account UUID or name", "in": "path", "name": "id", "required": true, "schema": {"type": "string"}},
          {"description": "History depth in days, 1 to 90 (default 30).", "in": "query", "name": "days", "required": false, "schema": {"format": "int32", "minimum": 0, "type": "integer"}}
        ],
        "responses": {
          "200": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProviderAccountUsageResponse"}}}, "description": "Usage of the account"},
          "401": {"description": "Unauthorized"},
          "403": {"description": "Forbidden"},
          "404": {"description": "Provider account not found"}
        },
        "security": [{"Bearer": []}],
        "summary": "Current windows of a Provider Account and their history. Admin only.",
        "tags": ["provider-accounts"]
      }
    },
    "/api/v1/api-keys": {
      "get": {
        "operationId": "list_api_keys",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ApiKeyResponse"
                  },
                  "type": "array"
                }
              }
            },
            "description": "List of API keys"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "List all API keys for the authenticated user.",
        "tags": [
          "api-keys"
        ]
      },
      "post": {
        "description": "# Errors\n\n- 400 if the name is empty or scopes are invalid",
        "operationId": "create_api_key",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApiKeyRequest"
              }
            }
          },
          "description": "API key configuration",
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateApiKeyResponse"
                }
              }
            },
            "description": "API key created successfully"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden (member trying to assign forbidden scopes)"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Create a new API key for the authenticated user.",
        "tags": [
          "api-keys"
        ]
      }
    },
    "/api/v1/api-keys/scopes": {
      "get": {
        "description": "Admins get all scopes, members only get read-only scopes.",
        "operationId": "available_scopes",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ScopeEntry"
                  },
                  "type": "array"
                }
              }
            },
            "description": "List of available scopes"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Return the list of scopes the current user is allowed to assign to API keys.",
        "tags": [
          "api-keys"
        ]
      }
    },
    "/api/v1/api-keys/{id}": {
      "delete": {
        "description": "# Errors\n\n- 404 if the key does not exist or belongs to another user",
        "operationId": "delete_api_key",
        "parameters": [
          {
            "description": "API key ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "API key deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "API key not found"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Delete an API key owned by the authenticated user.",
        "tags": [
          "api-keys"
        ]
      }
    },
...
    "/api/v1/approval-delegations/{id}": {
      "delete": {
        "description": "Only the delegator who granted it, or an admin, may revoke a delegation.\nExpired rows are still revocable: they are readable by ID even though they\nno longer appear in the list.\n\n# Errors\n\n- 401 if not authenticated\n- 403 if the caller is neither the delegator nor an admin\n- 404 if the delegation does not exist",
        "operationId": "delete_approval_delegation",
        "parameters": [
          {
            "description": "Delegation ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Delegation revoked"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Delegation not found"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Revoke an approval delegation.",
        "tags": [
          "approval-delegations"
        ]
      }
    },
...
    "/api/v1/auth/me": {
      "get": {
        "description": "# Errors\n\n- 401 if no valid token is provided\n- 404 if the user no longer exists in the store",
        "operationId": "me",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            },
            "description": "Current user profile"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Return the current user's profile.",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/password": {
      "patch": {
        "description": "Verifies the old password, hashes the new one, and persists it.\n\n# Errors\n\n- 400 if old password is incorrect or new password fails validation\n- 401 if no valid token is provided\n- 404 if the user no longer exists in the store",
        "operationId": "change_password",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          },
          "description": "Old and new passwords",
          "required": true
        },
        "responses": {
          "200": {
            "description": "Password changed successfully"
          },
          "400": {
            "description": "Invalid old password or validation error"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Change the authenticated user's password.",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/auth/sign-out": {
      "post": {
        "operationId": "sign_out",
        "responses": {
          "200": {
            "description": "User signed out successfully, cookies cleared"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Sign out the current user by clearing auth cookies.",
        "tags": [
          "auth"
        ]
      }
    },
...
  },
...
}